+++ demo-vue/src/vue-strickland/mapFormFieldValidationStates.js
